acm_open_decoder(): Use force_chans = 0 for "always trust header"
force_chans = 0 used to assume that plain ACM files are stereo (even if
their header says mono), as apparently that's the case for the files
from some games - but not all, Descent3 has both mono and stereo ACM
files, and their headers specify the correct number of channels, so
this quirk turns out to be a bad default, especially when there is no
way to tell libacm to trust the header.

Now force_chans = 0 assumes that the header is always correct, and the
old quirk mode (that assumes that plain ACM files are always stereo,
 but the headers of "WAVC" ACM files are always correct)
can be enabled with force_chans = -1.